pyproject.toml
[project]
name = "raggie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "ragie>=1.7.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.2",
]
